# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogoErrorResponse": {
        "additionalProperties": false,
        "description": "Error body for the 4 new catalog endpoints (US-038). Not retrofitted onto pre-existing operations. `indice` is present only for a nested-patrón validation failure on POST /api/categorias (CAT038-11).",
        "properties": {
          "code": {
            "type": "string"
          },
          "indice": {
            "type": "number"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "code"
        ],
        "type": "object"
      },
...
      "PatronResponse": {
        "additionalProperties": false,
        "description": "A classification pattern (US-038).",
        "properties": {
          "categoriaId": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "matchType": {
            "type": "string"
          },
          "patron": {
            "type": "string"
          },
          "prioridad": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "categoriaId",
          "patron",
          "matchType",
          "prioridad"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP contract for the MoneyDiary Express API, sourced from Zod schemas (ADR-011 amend).",
    "title": "MoneyDiary API",
    "version": "0.10.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/patrones": {
      "post": {
        "description": "Authenticated endpoint that creates a classification pattern under one of the caller's own categories (US-038, CAT038-05/06). Requires x-api-key + a valid session.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "categoriaId": {
                    "type": "string"
                  },
                  "matchType": {
                    "type": "string"
                  },
                  "patron": {
                    "type": "string"
                  },
                  "prioridad": {
                    "type": "number"
                  }
                },
                "required": [
                  "categoriaId",
                  "patron",
                  "matchType"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatronResponse"
                }
              }
            },
            "description": "Pattern created."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogoErrorResponse"
                }
              }
            },
            "description": "Invalid patron/matchType/prioridad, an invalid REGEX (write-time compile check), or a malformed request body."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogoErrorResponse"
                }
              }
            },
            "description": "Anti-enumeration: categoriaId does not exist or does not belong to the authenticated user."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogoErrorResponse"
                }
              }
            },
            "description": "A pattern with that text already exists for this user (case-insensitive)."
          }
        },
        "summary": "Create a classification pattern"
      }
    },
    "/api/patrones/{id}": {
      "delete": {
        "description": "Authenticated endpoint that deletes a pattern (US-038, CAT038-05/07). Requires x-api-key + a valid session.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Pattern deleted. No response body."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogoErrorResponse"
                }
              }
            },
            "description": "Anti-enumeration: the pattern does not exist or does not belong to the authenticated user."
          }
        },
        "summary": "Delete a classification pattern"
      },
      "patch": {
        "description": "Authenticated endpoint that partially updates a pattern (US-038, CAT038-05). `categoriaId` is NOT accepted — moving a pattern between categories is a non-goal. Requires x-api-key + a valid session.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "matchType": {
                    "type": "string"
                  },
                  "patron": {
                    "type": "string"
                  },
                  "prioridad": {
                    "type": "number"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatronResponse"
                }
              }
            },
            "description": "Pattern updated."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogoErrorResponse"
                }
              }
            },
            "description": "Invalid patron/matchType/prioridad, an invalid REGEX, an empty body, or a malformed request body."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogoErrorResponse"
                }
              }
            },
            "description": "Anti-enumeration: the pattern does not exist or does not belong to the authenticated user."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogoErrorResponse"
                }
              }
            },
            "description": "A pattern with that text already exists for this user (case-insensitive)."
          }
        },
        "summary": "Update a classification pattern"
      }
    },
...
  }
}